ADD: get, post for OwnerView and DogView

+++ dog_owner/dog/models.py
@@ -3,7 +3,7 @@ from django.db import models
 class Owner(models.Model):
     name = models.CharField(max_length=45)
     email = models.EmailField(max_length=254)
-    age = models.BigIntegerField
+    age = models.IntegerField
     class Meta: 
         db_table = 'owners' 
     def __str__(self):
